Relocate the framework test dir
