+++ src/plugins/data_source_management/public/components/frame_work_test/cache_intercept.test.ts
@@ -3,9 +3,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { SECURITY_DASHBOARDS_LOGOUT_URL } from '../../framework/constants';
-import { ASYNC_QUERY_ACCELERATIONS_CACHE, ASYNC_QUERY_DATASOURCE_CACHE } from '../utils/shared';
-import { catalogRequestIntercept } from './cache_intercept';
+import { SECURITY_DASHBOARDS_LOGOUT_URL } from '../../../framework/constants';
+import {
+  ASYNC_QUERY_ACCELERATIONS_CACHE,
+  ASYNC_QUERY_DATASOURCE_CACHE,
+} from '../../../framework/utils/shared';
+import { catalogRequestIntercept } from '../../../framework/catlog_cache/cache_intercept';
 
 interface LooseObject {
   [key: string]: any;
