+++ src/plugins/data_source_management/public/components/frame_work_test/utils/mocks.ts
@@ -12,7 +12,7 @@ import {
   DataSourceCacheData,
   DatasourceDetails,
   DatasourceType,
-} from '../../types';
+} from '../../../../framework/types';
 
 export const showDataConnectionsData = {
   schema: [
